backend/pyproject.toml
[project]
name = "voxscribe-backend"
version = "0.1.0"
description = "VoxScribe API Backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "pydantic-settings>=2.13.1",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]
